Make KeyListenerTracker more resilient
This adds a try/catch(Throwable) around all calls to KeyListener
methods that are inside a loop. Methods handling single KeyListeners
are not made more resilient, it makes sense to have the caller handle
errors in this case.

Also, reduce the amount of code duplication, and remove checks for
duplicate KeyListeners: addKeyListener ensures there are no such
duplicates.
As an additional layer of assurance, make the keyListeners member
private.